Add subaccount/org management guide
This is a first attempt at creating a guide for the new sub-account
capabilities and how they can be used to have multiple merchants.

I'm opening it to get a bit of feedback on it. I feel like I'm not that
good at writing this, but it serves as a good starting point I believe.
The images might need refresh since the PR is still in review.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>