Add workflow to trigger automatic client rebuild
We want to trigger a rebuild of our client libraries whenever the api
schema changes. For that purpose, create a workflow that detects changes
in the schema file and triggers another workflow on other repos. For
now, only the lune-ts repo is called. The target worflow produces a rebuild of
the library and open a PR if there are any changes.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>