Prevent infinite child parameter expansion
We have cases where a Parameter of oneOf presents a ToggableContent and
inside it's JSON representation. However, there were cases where this
JSON representation was calling parameter once again, leading to an
infinite loop scenario where child parameters could continously be
expanded.

Instead, make sure that the JSON representation does not toogle a
parameter if the JSON received is a parameter already resolved as a
oneOf.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>