First draft of documentation
This adds several fixes and improvements to previous version. Most
relevant:

- Improve how OpenAPI schema is processed. It now adds information about
 endpoints on the resources (for x-components) and adds a class to the
sidebar showcasing the endpoint method (get, post, delete etc.)
- Creation of JsonPropertyParser. This parses any property (object,
  string, allOf, oneOf etc.). The parsing is made iteratively to any
'children' of the property
- Creation of ParameterParser. This parses query/route parameters for
  endpoints.
- Creation of ResourceExample. This creates examples for any parsed
  property.
- Improvement and clean of component structure. We now have `Resource`
  and `Endpoint` as the main React Components, which use the created
parsers to present all the necessary information. Since `allOf` is not
yet ported to be displayed in a `allOf`, there's a soon to be deleted
`SchemaCombination` React component to process/showcase it.
- Several improvements to tsconfig, docusaurus.config.js, sidebar etc.

NOTE: `lune-ui-lib` seems currently in a weird place. I've been manually
linking it for now. Fixes to actual version will come later.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>