Improve URL (#5)
Instead of having a mix between camelCase and UpperCamelCase in the urls
when navigating, make the urls be `-` separated instead. Most of the PR
is about renaming the files to accomplish this, but some _category_.json
files required added to preserve the labels. This also required changes in the
index to point to the new urls and changes on the api-reference rebuild
script.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>