Access dashboard local storage (#3)
To allow our snippets of code to produce copy-pastable commands that
actually perform the operations, we access the value in the dashboard
local storage for a test API key. This is done by creating an IFrame and
posting a message to get a response back (in this case the test API key
stored in the dashboard local storage).

The test API key is then stored in the docs local storage and a callback
to refetch from the dashboard app is provided, currently only being
used on a test button that can be seen above the endpoint Snippet.

Since we can never know whenever test API key's change in the back,
this does mean that we might be showing a no longer valid API key, this
whoever is unavoidable since we don't want to maintain communication
with the dashboard nor DAPI from the docs. Better to provide a manual
mechanism to update when needed.

This requires: https://github.com/lune-climate/lune-dashboard/pull/665

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>