Fix examples (#59)
* Fix examples

We had several examples that had issues. After some digging, they all
originated from the same underlying issue. When parameters are passed
through, we weren't correctly detecting if they're a high level
resource. Since we now have this utility, we can simply use it. The
example was also being immediately solve as a rawExample which is wrong,
we might need to parse it further.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>

* Correctly populate $ref when parsing (#65)

For the high level components (direct markdown children), they do not
have the $ref property and instead rely on the presence of component.
To correctly populate it, we need to take that into account and populate
where necessary.

This was making the presentation bits be unable to distinguish between a
reference resource and a non reference one, which makes it impossible to
correctly present some cases.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>