Switch from Node 17 to Node 16
We should not be using Node 17.x, it reached its end of life 5 months
ago (on 2022-06-01[1]).

We can't upgrade to 18.x since Cloudflare Pages still doesn't support it
so the only solution we have is to technically downgrade to 16.x which
will be supported until 2023-09-11.

This resolves the problem of some of dependencies failing to upgrade to
newer versions (which versions refused to work with Node 17.x, like
[3]).

[1] https://github.com/nodejs/Release/tree/857d26d0514223ba0e17d54fdd1c85c03c2499c4#end-of-life-releases
[2] https://github.com/nodejs/Release/tree/857d26d0514223ba0e17d54fdd1c85c03c2499c4#release-schedule
[3] https://github.com/lune-climate/lune-docs/pull/4