Add support for multipart/form-data content type
We have just added this on the API, so we need to support it or build
fails (currently failing on the auto-update PR).

We now correctly process both content types and on the curl example act
according to this contentType. The final curl I believe is fully
functional, but I'm unable to curl since issues running BE locally.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>