Handle fully optional request body
With the introduction of the new PATCH endpoints, this means that the
request body can be fully optional. Since we only show examples on
required fields, this means we were showing 'undefined'. Instead,
simply show an empty object on this case if content type is json. In
case it's a multipart/form-data, we don't show the actual curl flag
to add data since it requires data (-F).

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>