Bring tracked files up to date.
Whenever we have changes to the API schema, we check the build is
successfull but the generated/edited/removed files of the build are not
saved to git. This means what is tracked on git and what is deployed is
quite different.

Sync things on git, building from current master.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>