Add TS library examples on endpoints
Showing examples of how one can use our Typescript library to use the
endpoints makes our library actually usefull, since atm there is no
direct connection between our documentation and our client.

For now, a snippet below the curl has been added to test and showcase,
but we probably want to move this to a tabbed snippet that allows to
either see the Curl example or the Lune-TS one.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>