Make the sidebar section labels more obvious
Following the principles of The Grand Unified Theory of Documentation[1]
and the conversation we had on the documention defficiencies[2] I
thought I'd go ahead and change the labels a little bit to make it
(more) clear what they are.

I haven't changed the directory names yet to avoid moving files around
and changing files' contents in the same commit.

[1] https://documentation.divio.com/
[2] https://linear.app/lune/issue/LUN-2365/docs-review