Handle referenced components that are references (#631)
Currently, we don't correctly handle the cases where a property is a
reference to a component that by itself is another reference. The result
is a build failure. We seem to have been avoiding this since usually we
don't need 2 levels of references. For the new analytics endpoints
though, this is the case.

The most high level properties are preserved, meaning that properties
like `description` are set to the component higher in the chain that
contains it.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>