Add identified flight methodology page
We want to document how our new identified flight methodology is done.
To make things visible and understandable a new page in the `Produce &
design guides` section was created.

Since this new page requires tables to be presented and the content is
gated, we require a new plugin for ReactMarkdown to deal with that since
ReactMarkdown follows the CommonMark standard which does not have
Tables. The plugin has been added to all Markdown content, allowing
footnotes, strikethrough, tables etc.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>