Add new changelog page
We want to introduce a changelog to the BE and use it in the docs. This
adds the capability for that, generating the page whenever we build from
the schema.

The current changelog is a dummy one to showcase the feature. Once [^1]
is in place we will be automatically syncing with the BE

[^1]: https://github.com/lune-climate/lune-backend/pull/5705

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>