Stop breaking the changelog Markdown syntax
The previous approach when processing the backend changelog entries was
to concatenate the lines of every day-specific changelog entry and put them
in a single bullet point within a section associated with that day.

That resulted in a few issues:

1. The lines were concatenated without any additional whitespace,
   therefore we ended up concatenating words together.
2. Theat approach was limited to very simple Markdown where.
   concatenating lines together could even work – as soon as we had any
   kind of code blocks or bullet lists this would break.
3. The approach introduces inflexibility in that we would have bullet
   points even in cases where we only had one change on a given day.
4. Empty lines were ignored in the concatenation process so that any
   kind of paragraph separation would be gone.

The issue 1 is what triggered this patch and the problem is visible in
this diff ("parameterenables" and such).

If we want all changelog daily entries to have bullet lists we can
modify the source changelog file to have that.