Sync changelog page with actual content
The content of the page is generated on build since it depends on the
changelog. We are having some issues where the automatic PR that updates
the changelog is not correctly adding these changes so the first step is
to sync the changes. A future PR will fix the underlying issue.

The changes introduced were added by simply running `yarn build`.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>