Update changelog page if changelog changes
We sync the changelog present in our backend repo via a github action.
This happens anytime a change to master occurs. This is the same
mechanism as to how we update the API schema.

This triggers a new workflow that creates the PR to automatically update
things in the docs. However, since the changelog page is generated via
the static changelog we want to also regenerate this page (done when
building) and add these changes automatically. This was not being done
which was making the page in the branch not be in sync with the actual
changelog, making building the project show changes in `master`.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>