Update changelog page if changelog changes (#1192)
We sync the changelog present in our backend repo via Github Actions.
This happens anytime a change to master occurs. This is the same
mechanism as to how we update the API schema.

This triggers a new workflow that creates the PR to automatically update
things in the docs. However, since the changelog page is generated via
the changelog `static/changelog.md` we want to also regenerate this
page (done when building) and add these changes automatically.

This was not being done which was making updates in the changelog
merged on the automatic PR's not update the changelog page. This
made the resulting code and the resulting release not match after the PR
was merged, since the actual changelog page code was not updated.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>