Improve changelog page (#1202)
After some debate during the demo in how to create a new API version, we
decided to make some changes to this page. The big rework is in how the
changelog events are presented: previous we grouped by date but it seems
grouping by API version makes it much more readable. But other small
changes/improvements were wanted.

List of changes:
- Group things by API version and not by date. Each API version then
  contains all the changes introduced with the dates
- Fix order to be cronologically reversed. This makes the latest changes
  sit at the top.
- Present dates in a different format than the calendar version one to
  avoid confusions. Went with `August 29, 2024`. This required adding
  `js-joda` as a development dependency.
- Due to change in how things are presented, the styling was adapted in
  order to be better. Do let me know what you think.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>