Fix automatic changelog update
This has been recently updated to include the changes in the changelog
page be automatically added when changes to the changelog are pushed.

However, issues exists: we weren't passing the github token required
for authentication nor correctly setting an identity which is a
requirement for committing changes to git.

This was making the process fail in CI with:

```
Author identity unknown

*** Please tell me who you are.
```

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>