Add smart scan guide
The new functionality is currently only shown to our customers via the
API schema. Since we have finished the first iteration of the feature we
want to document to our customers what it does and how they can use it.

I'm still a bit unsure how to best present this information, any
feedback is appreciated.

The example provided is probably not the best but again, I'm unsure of
what type of example we want to showcase to our customers, meaning
something that is expected to be how it's used and that also aligns very
well with our current emission factor data.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>