Add guide on how to create estimate using Mindee
We want to have a small guide for our customers in order for them to be
able to create a smart scan estimate using Mindee.

This is a how to guide that should give users like PM's and possibly
more technical users like developers to become able to use Mindee to
extract information from an invoice and create a smart scan estimate on
Lune based on that.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>